requirements.txt
Flask
python-dotenv
requests
selenium
webdriver-manager